Resolve: Correct paths afeter moving components
