+++ src/components/Card.js
@@ -1,4 +1,4 @@
-import "./App.css";
+// import "../src/App.css";
 import { Tag } from "./Tag.js";
 import { Description } from "./Description.js";
 import { Header } from "./Header.js";
