+++ src/components/Description.js
@@ -1,5 +1,5 @@
 import React from "react";
-import "./App.css";
+// import "../src/App.css";
 
 export const Description = () => {
   return (
