+++ src/components/Header.js
@@ -1,10 +1,6 @@
 import React from "react";
-import "./App.css";
+// import "../src/App.css";
 
 export const Header = () => {
-  return (
-    <header className="header">
-    Sports Gifts🎉
-  </header>
-  )
+  return <header className="header">Sports Gifts🎉</header>;
 };
