+++ src/components/cardsContainer.js
@@ -1,4 +1,4 @@
-import "./App.css";
+// import "../src/App.css";
 import {Card} from "./Card.js";
 // import { Tag } from "./Tag.js";
 // import { Description } from "./Description.js";
